MAINT: moved config routes to own file.
